+++ server/src/services/typescriptService/serviceHost.ts
@@ -1,7 +1,7 @@
 import * as path from 'path';
 import * as ts from 'typescript';
 import { URI } from 'vscode-uri';
-import { TextDocument } from 'vscode-languageserver-types';
+import { TextDocument } from 'vscode-languageserver-textdocument';
 import * as parseGitIgnore from 'parse-gitignore';
 
 import { LanguageModelCache } from '../../embeddedSupport/languageModelCache';